.NET 6 upgrade and UI project renamed to Demo
