+++ BlazorBootstrap.Demo/Pages/ConfirmDialogExamples.razor
