+++ BlazorBootstrap.Demo/Pages/Icons.razor
