+++ BlazorBootstrap.Demo/Pages/ToastsExamples.razor
