+++ BlazorBootstrap.Demo/Program.cs
@@ -1,0 +1,15 @@
+using BlazorBootstrap;
+using BlazorBootstrap.Demo;
+using Microsoft.AspNetCore.Components.Web;
+using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
+
+var builder = WebAssemblyHostBuilder.CreateDefault(args);
+builder.RootComponents.Add<App>("#app");
+builder.RootComponents.Add<HeadOutlet>("head::after");
+
+builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri(builder.HostEnvironment.BaseAddress) });
+
+// Inject services
+builder.Services.AddBlazorBootstrap();
+
+await builder.Build().RunAsync();
