+++ BlazorBootstrap.Demo/Properties/launchSettings.json
@@ -8,7 +8,7 @@
     }
   },
   "profiles": {
-    "BlazorBootstrap.UI": {
+    "BlazorBootstrap.Demo": {
       "commandName": "Project",
       "dotnetRunMessages": "true",
       "launchBrowser": true,
