+++ BlazorBootstrap.Demo/Shared/NavMenu.razor
