+++ BlazorBootstrap.Demo/_Imports.razor
@@ -6,5 +6,5 @@
 @using Microsoft.AspNetCore.Components.Web.Virtualization
 @using Microsoft.AspNetCore.Components.WebAssembly.Http
 @using Microsoft.JSInterop
-@using BlazorBootstrap.UI
-@using BlazorBootstrap.UI.Shared
+@using BlazorBootstrap.Demo
+@using BlazorBootstrap.Demo.Shared
