--- BlazorBootstrap.UI/Program.cs
@@ -1,22 +1,0 @@
-
-using Microsoft.AspNetCore.Components.Web;
-using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
-
-namespace BlazorBootstrap.UI
-{
-    public class Program
-    {
-        public static async Task Main(string[] args)
-        {
-            var builder = WebAssemblyHostBuilder.CreateDefault(args);
-            builder.RootComponents.Add<App>("#app");
-
-            builder.Services.AddScoped(sp => new HttpClient { BaseAddress = new Uri(builder.HostEnvironment.BaseAddress) });
-
-            // Inject services           
-            builder.Services.AddBlazorBootstrap();
-
-            await builder.Build().RunAsync();
-        }
-    }
-}
